test the IP address and subnet generation
Test the IP address and subnet generation to avoid breaking the existing
generation when updating or refactoring it. Given that the subnet
generation uses some binary operations, it is hard to spot potential
errors.

For this usecase the function generating the subnet was externalized to
be accessible from the test code instead of an inlined function.